osd: fix parsing invalid json produced by Ceph after 10.2.5 when OSD is out
Due to http://tracker.ceph.com/issues/18239 Ceph will generate invalid
json output when there is any of OSD out of the cluster. Since it is
invalid, golang is unable to unmarshal the json and can result in
confusing metrics. Replacing -nan to 0 is certainly a workaround,
but this can temporarily prevent wrong metrics.

Fixes: https://github.com/digitalocean/ceph_exporter/issues/48