pool: remove error log when no pools present
If there's no pools created yet (which can be true after a ceph
cluster is initially created, but before any pools are present),
there'll be a log every minute "[ERROR] failed collecting pool
usage metrics: no pools found in the cluster to report stats on".

IMO that's not actually an error, it's just the way things are,
so doesn't really need to be reported.

Signed-off-by: Tim Serong <tserong@suse.com>