feat: add osd full/nearfull/backfillfull ratio
Add new gauge to show osd full/nearfull/backfillfull ratio.
Not only do we need to know whether a osd is full or not, we
also want to know the exact full ratio was for a cluster.
For many clusters which have different full ratio set, this
should be meaningfull.

Signed-off-by: haoyixing <haoyixing@kuaishou.com>